Major reorg of rvic package.
- Moved rvic source files to rvic core package
- Create rvic_model script to be installed in bin/ (usage: rvic_model (rvic_function) config_file ...)
- Added setup.py and travis info
