+++ .travis.yml
@@ -1,0 +1,24 @@
+language: python
+
+python:
+  - "2.7"
+# Setup anaconda
+before_install:
+  - wget http://repo.continuum.io/miniconda/Miniconda-2.2.2-Linux-x86_64.sh -O miniconda.sh
+  - chmod +x miniconda.sh
+  - ./miniconda.sh -b
+  - export PATH=/home/travis/anaconda/bin:$PATH
+  # Update conda itself
+  - conda update --yes conda
+  # The next couple lines fix a crash with multiprocessing on Travis and are not specific to using Miniconda
+  - sudo rm -rf /dev/shm
+  - sudo ln -s /run/shm /dev/shm
+# Install packages
+install:
+  - conda create --yes -n rvic_env python=$TRAVIS_PYTHON_VERSION 
+  - source activate rvic_env
+  - pip install -r ci/requireemnts-2.7.txt
+  - python setup.py install
+# Run test
+script:
+  - py.test
