+++ ci/requirements-2.7-dev.txt
@@ -1,0 +1,6 @@
+cdo==1.2.3
+matplotlib==1.3.1
+netCDF4==1.0.8
+numpy==1.8.1
+scipy==0.13.3
+pytest==2.5.2
