+++ ci/requirements-2.7.txt
@@ -1,0 +1,4 @@
+matplotlib==1.3.1
+netCDF4==1.0.8
+numpy==1.8.1
+scipy==0.13.3
