+++ rvic/core/aggregate.py
