+++ rvic/core/utilities.py
