+++ rvic/core/variables.py
